add Builder methods for public API, part 3.
