+++ cdm/core/src/main/java/ucar/nc2/iosp/netcdf3/N3header.java
@@ -23,8 +23,7 @@ public class N3header extends NCheader {
   private static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(N3header.class);
 
   static final byte[] MAGIC = {0x43, 0x44, 0x46, 0x01};
-  private static final long MAX_UNSIGNED_INT = 0x00000000ffffffffL;
-
+  static final long MAX_UNSIGNED_INT = 0x00000000ffffffffL;
   static final byte[] MAGIC_LONG = {0x43, 0x44, 0x46, 0x02}; // 64-bit offset format : only affects the
                                                              // variable offset value
   static final int MAGIC_DIM = 10;
@@ -45,15 +44,29 @@ public class N3header extends NCheader {
   public static boolean disallowFileTruncation; // see NetcdfFile.setDebugFlags
   public static boolean debugHeaderSize; // see NetcdfFile.setDebugFlags
 
-  private static boolean debugVariablePos;
-  private static boolean debugStreaming;
+  private static boolean debugVariablePos = false;
+  private static boolean debugStreaming = false;
+
+
+  // variable info for reading/writing
+  static class Vinfo {
+    long vsize; // size of array in bytes. if isRecord, size per record.
+    long begin; // offset of start of data from start of file
+    boolean isRecord; // is it a record variable?
+    long attsPos; // attributes start here - used for update
+
+    Vinfo(long vsize, long begin, boolean isRecord, long attsPos) {
+      this.vsize = vsize;
+      this.begin = begin;
+      this.isRecord = isRecord;
+      this.attsPos = attsPos;
+    }
+  }
 
   ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
   ucar.unidata.io.RandomAccessFile raf;
-  private ucar.nc2.NetcdfFile ncfile;
-  private List<Variable> uvars = new ArrayList<>(); // vars that have the unlimited dimension
-  private Dimension udim; // the unlimited dimension
+  ucar.nc2.NetcdfFile ncfile;
 
   // N3iosp needs access to these
   boolean isStreaming; // is streaming (numrecs = -1)
@@ -63,16 +76,16 @@ public class N3header extends NCheader {
 
   private boolean useLongOffset;
   private long nonRecordDataSize; // size of non-record variables
-  private long dataStart = Long.MAX_VALUE; // where the data starts
+  private Dimension udim; // the unlimited dimension
 
-  private long globalAttsPos; // global attributes start here - used for update
+  long dataStart = Long.MAX_VALUE; // where the data starts
 
   /*
-   * Notes
-   * - dimensions are signed or unsigned ? in java, must be signed, so are limited to 2^31, not 2^32
-   * " Each fixed-size variable and the data for one record's worth of a single record variable are limited in size to a
-   * little less
-   * that 4 GiB, which is twice the size limit in versions earlier than netCDF 3.6."
+   * Notes:
+   * In netcdf-3 are dimensions signed or unsigned?
+   * In java, integers are signed, so are limited to 2^31, not 2^32
+   * "Each fixed-size variable and the data for one record's worth of a single record variable are limited in size to a
+   * little less than 4 GiB, which is twice the size limit in versions earlier than netCDF 3.6."
    */
 
   /**
@@ -83,9 +96,23 @@ public class N3header extends NCheader {
    * @param fout optional for debug message, may be null
    * @throws IOException on read error
    */
-  void read(ucar.unidata.io.RandomAccessFile raf, ucar.nc2.NetcdfFile ncfile, Formatter fout) throws IOException {
-    this.raf = raf;
+  void read(RandomAccessFile raf, NetcdfFile ncfile, Formatter fout) throws IOException {
     this.ncfile = ncfile;
+    Group.Builder root = Group.builder().setNcfile(ncfile);
+    read(raf, root, fout);
+    // LOOK ncfile.setRootGroup(root.build());
+  }
+
+    /**
+     * Read the header and populate the ncfile
+     *
+     * @param raf read from this file
+     * @param root the root Group builder to populate.
+     * @param debugOut optional for debug message, may be null
+     * @throws IOException on read error
+     */
+  private void read(ucar.unidata.io.RandomAccessFile raf, Group.Builder root, Formatter debugOut) throws IOException {
+    this.raf = raf;
 
     long actualSize = raf.length();
     nonRecordDataSize = 0; // length of non-record data
@@ -108,8 +135,8 @@ public class N3header extends NCheader {
 
     // number of records
     numrecs = raf.readInt();
-    if (fout != null)
-      fout.format("numrecs= %d%n", numrecs);
+    if (debugOut != null)
+      debugOut.format("numrecs= %d%n", numrecs);
     if (numrecs == -1) {
       isStreaming = true;
       numrecs = 0;
@@ -124,31 +151,32 @@ public class N3header extends NCheader {
       if (magic != MAGIC_DIM)
         throw new IOException("Misformed netCDF file - dim magic number wrong " + raf.getLocation());
       numdims = raf.readInt();
-      if (fout != null)
-        fout.format("numdims= %d%n", numdims);
+      if (debugOut != null)
+        debugOut.format("numdims= %d%n", numdims);
     }
 
+    // Must keep dimensions in strict order
+    List<Variable.Builder> uvars = new ArrayList<>(); // vars that have the unlimited dimension
+    ArrayList<Dimension> fileDimensions = new ArrayList<>();
     for (int i = 0; i < numdims; i++) {
-      if (fout != null)
-        fout.format("  dim %d pos= %d%n", i, raf.getFilePointer());
+      if (debugOut != null)
+        debugOut.format("  dim %d pos= %d%n", i, raf.getFilePointer());
       String name = readString();
       int len = raf.readInt();
       Dimension dim;
       if (len == 0) {
-        dim = new Dimension(name, numrecs, true, true, false);
+        dim = Dimension.builder(name, numrecs).setIsUnlimited(true).build();
         udim = dim;
       } else {
-        dim = new Dimension(name, len, true, false, false);
+        dim = Dimension.builder(name, numrecs).build();
       }
-
-      ncfile.addDimension(null, dim);
-      if (fout != null)
-        fout.format(" added dimension %s%n", dim);
+      fileDimensions.add(dim);
+      if (debugOut != null)
+        debugOut.format(" added dimension %s%n", dim);
     }
 
     // global attributes
-    globalAttsPos = raf.getFilePointer();
-    readAtts(ncfile.getRootGroup(), fout);
+    readAtts(root.getAttributeContainer(), debugOut);
 
     // variables
     int nvars = 0;
@@ -159,17 +187,17 @@ public class N3header extends NCheader {
       if (magic != MAGIC_VAR)
         throw new IOException("Misformed netCDF file  - var magic number wrong " + raf.getLocation());
       nvars = raf.readInt();
-      if (fout != null)
-        fout.format("numdims= %d%n", numdims);
+      if (debugOut != null)
+        debugOut.format("numdims= %d%n", numdims);
     }
-    if (fout != null)
-      fout.format("num variables= %d%n", nvars);
+    if (debugOut != null)
+      debugOut.format("num variables= %d%n", nvars);
 
     // loop over variables
     for (int i = 0; i < nvars; i++) {
       long startPos = raf.getFilePointer();
       String name = readString();
-      Variable var = new Variable(ncfile, ncfile.getRootGroup(), null, name);
+      Variable.Builder var = Variable.builder().setName(name);
 
       // get element count in non-record dimensions
       long velems = 1;
@@ -178,7 +206,7 @@ public class N3header extends NCheader {
       List<Dimension> dims = new ArrayList<>();
       for (int j = 0; j < rank; j++) {
         int dimIndex = raf.readInt();
-        Dimension dim = ncfile.getRootGroup().getDimensions().get(dimIndex); // note relies on ordering
+        Dimension dim = fileDimensions.get(dimIndex);
         if (dim.isUnlimited()) {
           isRecord = true;
           uvars.add(var); // track record variables
@@ -187,18 +215,18 @@ public class N3header extends NCheader {
 
         dims.add(dim);
       }
-      var.setDimensions(dims);
+      var.addDimensions(dims);
 
-      if (fout != null) {
-        fout.format("---name=<%s> dims = [", name);
+      if (debugOut != null) {
+        debugOut.format("---name=<%s> dims = [", name);
         for (Dimension dim : dims)
-          fout.format("%s ", dim.getShortName());
-        fout.format("]%n");
+          debugOut.format("%s ", dim.getShortName());
+        debugOut.format("]%n");
       }
 
       // variable attributes
       long varAttsPos = raf.getFilePointer();
-      readAtts(var, fout);
+      readAtts(var.getAttributeContainer(), debugOut);
 
       // data type
       int type = raf.readInt();
@@ -206,15 +234,15 @@ public class N3header extends NCheader {
       var.setDataType(dataType);
 
       // size and beginning data position in file
-      long vsize = (long) raf.readInt();
+      long vsize = raf.readInt();
       long begin = useLongOffset ? raf.readLong() : (long) raf.readInt();
 
-      if (fout != null) {
-        fout.format(" name= %s type=%d vsize=%s velems=%d begin= %d isRecord=%s attsPos=%d%n", name, type, vsize,
+      if (debugOut != null) {
+        debugOut.format(" name= %s type=%d vsize=%s velems=%d begin= %d isRecord=%s attsPos=%d%n", name, type, vsize,
             velems, begin, isRecord, varAttsPos);
         long calcVsize = (velems + padding(velems)) * dataType.getSize();
         if (vsize != calcVsize)
-          fout.format(" *** readVsize %d != calcVsize %d%n", vsize, calcVsize);
+          debugOut.format(" *** readVsize %d != calcVsize %d%n", vsize, calcVsize);
       }
       if (vsize < 0) {
         vsize = (velems + padding(velems)) * dataType.getSize();
@@ -231,18 +259,7 @@ public class N3header extends NCheader {
       }
 
       dataStart = Math.min(dataStart, begin);
-
-      if (debugVariablePos)
-        System.out.printf("%s begin at=%d end=%d  isRecord=%s nonRecordDataSize=%d%n", var.getFullName(), begin,
-            (begin + vsize), isRecord, nonRecordDataSize);
-      if (fout != null)
-        fout.format("%s begin at=%d end=%d  isRecord=%s nonRecordDataSize=%d%n", var.getFullName(), begin,
-            (begin + vsize), isRecord, nonRecordDataSize);
-      if (debugHeaderSize)
-        System.out.printf("%s header size=%d data size= %d%n", var.getFullName(), (raf.getFilePointer() - startPos),
-            vsize);
-
-      ncfile.addVariable(null, var);
+      root.addVariable(var);
     }
 
     pos = raf.getFilePointer();
@@ -257,7 +274,7 @@ public class N3header extends NCheader {
     if (uvars.isEmpty()) // if there are no record variables
       recStart = 0;
 
-    // Check if file affected by bug CDM-52 (netCDF-Java library used incorrect padding when
+    /* LOOK Check if file affected by bug CDM-52 (netCDF-Java library used incorrect padding when
     // the file contained only one record variable and it was of type byte, char, or short).
     if (uvars.size() == 1) {
       Variable uvar = uvars.get(0);
@@ -276,7 +293,7 @@ public class N3header extends NCheader {
           vinfo.vsize = vsize;
         }
       }
-    }
+    } */
 
     if (debugHeaderSize) {
       System.out.println("  filePointer = " + pos + " dataStart=" + dataStart);
@@ -288,7 +305,7 @@ public class N3header extends NCheader {
       System.out.println("  actualSize= " + actualSize);
     }
 
-    // check for streaming file - numrecs must be calculated
+    /* LOOK check for streaming file - numrecs must be calculated
     if (isStreaming) {
       long recordSpace = actualSize - recStart;
       numrecs = recsize == 0 ? 0 : (int) (recordSpace / recsize);
@@ -306,7 +323,7 @@ public class N3header extends NCheader {
           uvar.invalidateCache();
         }
       }
-    }
+    } */
 
     // check for truncated files
     // theres a "wart" that allows a file to be up to 3 bytes smaller than you expect.
@@ -358,6 +375,7 @@ public class N3header extends NCheader {
     }
   }
 
+  /* LOOK
   synchronized boolean removeRecordStructure() {
     boolean found = false;
     for (Variable v : uvars) {
@@ -376,8 +394,8 @@ public class N3header extends NCheader {
   synchronized boolean makeRecordStructure() {
     // create record structure
     if (!uvars.isEmpty()) {
-      Structure recordStructure = new Structure(ncfile, ncfile.getRootGroup(), null, "record");
-      recordStructure.setDimensions(udim.getShortName());
+      Structure.Builder builder = Structure.builder().setName("record").setNcfile(ncfile).setParent(ncfile.getRootGroup());
+      builder.setDimensions(udim.getShortName());
       for (Variable v : uvars) {
         Variable memberV;
         try {
@@ -387,14 +405,7 @@ public class N3header extends NCheader {
           return false;
         }
         memberV.setParentStructure(recordStructure);
-        // memberV.createNewCache(); // decouple caching - could use this ??
-
-        // remove record dimension
-        // List<Dimension> dims = new ArrayList<Dimension>(v.getDimensions());
-        // dims.remove(0);
-        // memberV.setDimensions(dims);
-
-        recordStructure.addMemberVariable(memberV);
+        builder.addMemberVariable(memberV);
       }
 
       uvars.add(recordStructure);
@@ -404,8 +415,7 @@ public class N3header extends NCheader {
     }
 
     return false;
-  }
-
+  } */
 
   private int readAtts(AttributeContainer atts, Formatter fout) throws IOException {
     int natts = 0;
@@ -435,7 +445,7 @@ public class N3header extends NCheader {
           val = "";
         if (fout != null)
           fout.format(" end read String val pos= %d%n", raf.getFilePointer());
-        att = new Attribute(name, val);
+        att = Attribute.builder(name).setStringValue(val).build();
 
       } else {
         if (fout != null)
@@ -443,11 +453,9 @@ public class N3header extends NCheader {
         int nelems = raf.readInt();
 
         DataType dtype = getDataType(type);
+        Attribute.Builder builder = Attribute.builder(name).setDataType(dtype);
 
-        if (nelems == 0) {
-          att = new Attribute(name, dtype); // empty - no values
-
-        } else {
+        if (nelems > 0) {
           int[] shape = {nelems};
           Array arr = Array.factory(dtype, shape);
           IndexIterator ii = arr.getIndexIterator();
@@ -455,9 +463,10 @@ public class N3header extends NCheader {
           for (int j = 0; j < nelems; j++)
             nbytes += readAttributeValue(dtype, ii);
 
-          att = new Attribute(name, arr); // no validation !!
+          builder.setValues(arr); // no validation !!
           skip(nbytes);
         }
+        att = builder.build();
 
         if (fout != null)
           fout.format(" end read val pos= %d%n", raf.getFilePointer());
@@ -555,7 +564,7 @@ public class N3header extends NCheader {
     return pad;
   }
 
-  private DataType getDataType(int type) {
+  static DataType getDataType(int type) {
     switch (type) {
       case 1:
         return DataType.BYTE;
@@ -590,528 +599,4 @@ public class N3header extends NCheader {
 
     throw new IllegalArgumentException("unknown DataType == " + dt);
   }
-
-  /**
-   * Write the header out, based on ncfile structures.
-   *
-   * @param raf write to this file
-   * @param ncfile the header of this NetcdfFile
-   * @param extra if > 0, pad header with extra bytes
-   * @param largeFile if large file format
-   * @param fout debugging output sent to here
-   * @throws IOException on write error
-   */
-  void create(ucar.unidata.io.RandomAccessFile raf, ucar.nc2.NetcdfFile ncfile, int extra, boolean largeFile,
-      Formatter fout) throws IOException {
-    this.raf = raf;
-    this.ncfile = ncfile;
-
-    writeHeader(extra, largeFile, false, fout);
-  }
-
-  /**
-   * Sneaky way to make the header bigger, if there is room for it
-   *
-   * @param largeFile is large file format
-   * @param fout put debug messages here, mnay be null
-   * @return true if it worked
-   */
-  boolean rewriteHeader(boolean largeFile, Formatter fout) throws IOException {
-    int want = sizeHeader(largeFile);
-    if (want > dataStart)
-      return false;
-
-    writeHeader(0, largeFile, true, fout);
-    return true;
-  }
-
-  void writeHeader(int extra, boolean largeFile, boolean keepDataStart, Formatter fout) throws IOException {
-    this.useLongOffset = largeFile;
-    nonRecordDataSize = 0; // length of non-record data
-    recsize = 0; // length of single record
-    recStart = Long.MAX_VALUE; // where the record data starts
-
-    // magic number
-    raf.seek(0);
-    raf.write(largeFile ? N3header.MAGIC_LONG : N3header.MAGIC);
-
-    // numrecs
-    raf.writeInt(0);
-
-    // dims
-    List<Dimension> dims = ncfile.getDimensions();
-    int numdims = dims.size();
-    if (numdims == 0) {
-      raf.writeInt(0);
-      raf.writeInt(0);
-    } else {
-      raf.writeInt(N3header.MAGIC_DIM);
-      raf.writeInt(numdims);
-    }
-    for (int i = 0; i < numdims; i++) {
-      Dimension dim = dims.get(i);
-      if (fout != null)
-        fout.format("  dim %d pos %d%n", i, raf.getFilePointer());
-      writeString(dim.getShortName());
-      raf.writeInt(dim.isUnlimited() ? 0 : dim.getLength());
-      if (dim.isUnlimited())
-        udim = dim;
-    }
-
-    // global attributes
-    globalAttsPos = raf.getFilePointer();
-    writeAtts(ncfile.getGlobalAttributes(), fout);
-
-    // variables
-    List<Variable> vars = ncfile.getVariables();
-
-    // Track record variables.
-    for (Variable curVar : vars) {
-      if (curVar.isUnlimited()) {
-        uvars.add(curVar);
-      }
-    }
-    writeVars(vars, largeFile, fout);
-
-    // now calculate where things go
-    if (!keepDataStart) {
-      dataStart = raf.getFilePointer();
-      if (extra > 0)
-        dataStart += extra;
-    }
-    long pos = dataStart;
-
-    // non-record variable starting positions
-    for (Variable var : vars) {
-      Vinfo vinfo = (Vinfo) var.getSPobject();
-      if (!vinfo.isRecord) {
-        raf.seek(vinfo.begin);
-
-        if (largeFile)
-          raf.writeLong(pos);
-        else {
-          if (pos > Integer.MAX_VALUE)
-            throw new IllegalArgumentException("Variable starting pos=" + pos + " may not exceed " + Integer.MAX_VALUE);
-          raf.writeInt((int) pos);
-        }
-
-        vinfo.begin = pos;
-        if (fout != null)
-          fout.format("  %s begin at = %d end= %d%n", var.getFullName(), vinfo.begin, (vinfo.begin + vinfo.vsize));
-        pos += vinfo.vsize;
-
-        // track how big each record is
-        nonRecordDataSize = Math.max(nonRecordDataSize, vinfo.begin + vinfo.vsize);
-      }
-    }
-
-    recStart = pos; // record variables start here
-
-    // record variable starting positions
-    for (Variable var : vars) {
-      Vinfo vinfo = (Vinfo) var.getSPobject();
-      if (vinfo.isRecord) {
-        raf.seek(vinfo.begin);
-
-        if (largeFile)
-          raf.writeLong(pos);
-        else
-          raf.writeInt((int) pos);
-
-        vinfo.begin = pos;
-        if (fout != null)
-          fout.format(" %s record begin at = %d%n", var.getFullName(), dataStart);
-        pos += vinfo.vsize;
-
-        // track how big each record is
-        recsize += vinfo.vsize;
-        recStart = Math.min(recStart, vinfo.begin);
-      }
-    }
-
-    if (nonRecordDataSize > 0) // if there are non-record variables
-      nonRecordDataSize -= dataStart;
-    if (uvars.isEmpty()) // if there are no record variables
-      recStart = 0;
-  }
-
-  // calculate the size writing a header would take
-  int sizeHeader(boolean largeFile) {
-    int size = 4; // magic number
-    size += 4; // numrecs
-
-    // dims
-    size += 8; // magic, ndims
-    for (Dimension dim : ncfile.getDimensions())
-      size += sizeString(dim.getShortName()) + 4; // name, len
-
-    // global attributes
-    size += sizeAtts(ncfile.getGlobalAttributes());
-
-    // variables
-    size += 8; // magic, nvars
-    for (Variable var : ncfile.getVariables()) {
-      size += sizeString(var.getShortName());
-
-      // dimensions
-      size += 4; // ndims
-      size += 4 * var.getDimensions().size(); // dim id
-
-      // variable attributes
-      size += sizeAtts(var.getAttributes());
-
-      size += 8; // data type, variable size
-      size += (largeFile) ? 8 : 4;
-    }
-
-    return size;
-  }
-
-  private void writeAtts(List<Attribute> atts, Formatter fout) throws IOException {
-
-    int n = atts.size();
-    if (n == 0) {
-      raf.writeInt(0);
-      raf.writeInt(0);
-    } else {
-      raf.writeInt(MAGIC_ATT);
-      raf.writeInt(n);
-    }
-
-    for (int i = 0; i < n; i++) {
-      if (fout != null)
-        fout.format("***att %d pos= %d%n", i, raf.getFilePointer());
-      Attribute att = atts.get(i);
-
-      writeString(att.getShortName());
-      int type = getType(att.getDataType());
-      raf.writeInt(type);
-
-      if (type == 2) {
-        writeStringValues(att);
-      } else {
-        int nelems = att.getLength();
-        raf.writeInt(nelems);
-        int nbytes = 0;
-        for (int j = 0; j < nelems; j++)
-          nbytes += writeAttributeValue(att.getNumericValue(j));
-        pad(nbytes, (byte) 0);
-        if (fout != null)
-          fout.format(" end write val pos= %d%n", raf.getFilePointer());
-      }
-      if (fout != null)
-        fout.format("  %s%n", att);
-    }
-  }
-
-  private int sizeAtts(List<Attribute> atts) {
-    int size = 8; // magic, natts
-
-    for (Attribute att : atts) {
-      size += sizeString(att.getShortName());
-      size += 4; // type
-
-      int type = getType(att.getDataType());
-      if (type == 2) {
-        size += sizeStringValues(att);
-      } else {
-        size += 4; // nelems
-        int nelems = att.getLength();
-        int nbytes = 0;
-        for (int j = 0; j < nelems; j++)
-          nbytes += sizeAttributeValue(att.getNumericValue(j));
-        size += nbytes;
-        size += padding(nbytes);
-      }
-    }
-    return size;
-  }
-
-  private void writeStringValues(Attribute att) throws IOException {
-    int n = att.getLength();
-    if (n == 1)
-      writeString(att.getStringValue());
-    else {
-      StringBuilder values = new StringBuilder();
-      for (int i = 0; i < n; i++)
-        values.append(att.getStringValue(i));
-      writeString(values.toString());
-    }
-  }
-
-  private int sizeStringValues(Attribute att) {
-    int size = 0;
-    int n = att.getLength();
-    if (n == 1)
-      size += sizeString(att.getStringValue());
-    else {
-      StringBuilder values = new StringBuilder();
-      for (int i = 0; i < n; i++)
-        values.append(att.getStringValue(i));
-      size += sizeString(values.toString());
-    }
-
-    return size;
-  }
-
-  private int writeAttributeValue(Number numValue) throws IOException {
-    if (numValue instanceof Byte) {
-      raf.write(numValue.byteValue());
-      return 1;
-
-    } else if (numValue instanceof Short) {
-      raf.writeShort(numValue.shortValue());
-      return 2;
-
-    } else if (numValue instanceof Integer) {
-      raf.writeInt(numValue.intValue());
-      return 4;
-
-    } else if (numValue instanceof Float) {
-      raf.writeFloat(numValue.floatValue());
-      return 4;
-
-    } else if (numValue instanceof Double) {
-      raf.writeDouble(numValue.doubleValue());
-      return 8;
-    }
-
-    throw new IllegalStateException("unknown attribute type == " + numValue.getClass().getName());
-  }
-
-  private int sizeAttributeValue(Number numValue) {
-
-    if (numValue instanceof Byte) {
-      return 1;
-
-    } else if (numValue instanceof Short) {
-      return 2;
-
-    } else if (numValue instanceof Integer) {
-      return 4;
-
-    } else if (numValue instanceof Float) {
-      return 4;
-
-    } else if (numValue instanceof Double) {
-      return 8;
-    }
-
-    throw new IllegalStateException("unknown attribute type == " + numValue.getClass().getName());
-  }
-
-  private void writeVars(List<Variable> vars, boolean largeFile, Formatter fout) throws IOException {
-    int n = vars.size();
-    if (n == 0) {
-      raf.writeInt(0);
-      raf.writeInt(0);
-    } else {
-      raf.writeInt(MAGIC_VAR);
-      raf.writeInt(n);
-    }
-
-    for (Variable var : vars) {
-      writeString(var.getShortName());
-
-      // dimensions
-      long vsize = var.getDataType().getSize(); // works for all netcdf-3 data types
-      List<Dimension> dims = var.getDimensions();
-      raf.writeInt(dims.size());
-      for (Dimension dim : dims) {
-        int dimIndex = findDimensionIndex(ncfile, dim);
-        raf.writeInt(dimIndex);
-
-        if (!dim.isUnlimited())
-          vsize *= dim.getLength();
-      }
-      long unpaddedVsize = vsize;
-      vsize += padding(vsize);
-
-      // variable attributes
-      long varAttsPos = raf.getFilePointer();
-      writeAtts(var.getAttributes(), fout);
-
-      // data type, variable size, beginning file position
-      DataType dtype = var.getDataType();
-      int type = getType(dtype);
-      raf.writeInt(type);
-
-      int vsizeWrite = (vsize < MAX_UNSIGNED_INT) ? (int) vsize : -1;
-      raf.writeInt(vsizeWrite);
-      long pos = raf.getFilePointer();
-      if (largeFile)
-        raf.writeLong(0); // come back to this later
-      else
-        raf.writeInt(0); // come back to this later
-
-      // From nc3 file format specification
-      // (http://www.unidata.ucar.edu/software/netcdf/docs/netcdf.html#NetCDF-Classic-Format):
-      // Note on padding: In the special case of only a single record variable of character,
-      // byte, or short type, no padding is used between data values.
-      // 2/15/2011: we will continue to write the (incorrect) padded vsize into the header, but we will use the unpadded
-      // size to read/write
-      if (uvars.size() == 1 && uvars.get(0) == var)
-        if ((dtype == DataType.CHAR) || (dtype == DataType.BYTE) || (dtype == DataType.SHORT))
-          vsize = unpaddedVsize;
-
-      var.setSPobject(new Vinfo(vsize, pos, var.isUnlimited(), varAttsPos));
-    }
-  }
-
-  // write a string then pad to 4 byte boundary
-  private void writeString(String s) throws IOException {
-    byte[] b = s.getBytes(StandardCharsets.UTF_8); // all strings are encoded in UTF-8 Unicode.
-    raf.writeInt(b.length);
-    raf.write(b);
-    pad(b.length, (byte) 0);
-  }
-
-  private int sizeString(String s) {
-    int size = s.length() + 4;
-    return size + padding(s.length());
-  }
-
-  private int findDimensionIndex(NetcdfFile ncfile, Dimension wantDim) {
-    List<Dimension> dims = ncfile.getDimensions();
-    for (int i = 0; i < dims.size(); i++) {
-      Dimension dim = dims.get(i);
-      if (dim.equals(wantDim))
-        return i;
-    }
-    throw new IllegalStateException("unknown Dimension == " + wantDim);
-  }
-
-  // pad to a 4 byte boundary
-  private void pad(int nbytes, byte fill) throws IOException {
-    int pad = padding(nbytes);
-    for (int i = 0; i < pad; i++)
-      raf.write(fill);
-  }
-
-  void writeNumrecs() throws IOException {
-    // set number of records in the header
-    raf.seek(4);
-    raf.writeInt(numrecs);
-  }
-
-  void setNumrecs(int n) {
-    this.numrecs = n;
-  }
-
-  synchronized boolean synchNumrecs() throws IOException {
-    // check number of records in the header
-    // gotta bypass the RAF buffer
-    int n = raf.readIntUnbuffered(4);
-    if (n == this.numrecs)
-      return false;
-    if (n < 0) // streaming
-      return false;
-
-    // update everything
-    this.numrecs = n;
-
-    // set it in the unlimited dimension
-    udim.setLength(this.numrecs);
-
-    // set it in all of the record variables
-    for (Variable uvar : uvars) {
-      uvar.resetShape();
-      uvar.invalidateCache();
-    }
-
-    return true;
-  }
-
-
-  // variable info for reading/writing
-  static class Vinfo {
-    long vsize; // size of array in bytes. if isRecord, size per record.
-    long begin; // offset of start of data from start of file
-    boolean isRecord; // is it a record variable?
-    long attsPos; // attributes start here - used for update
-
-    Vinfo(long vsize, long begin, boolean isRecord, long attsPos) {
-      this.vsize = vsize;
-      this.begin = begin;
-      this.isRecord = isRecord;
-      this.attsPos = attsPos;
-    }
-  }
-
-  ///////////////////////////////////////////////////////////////////////////////
-  // tricky - perhaps deprecate ?
-
-  void updateAttribute(ucar.nc2.Variable v2, Attribute att) throws IOException {
-    long pos;
-    if (v2 == null)
-      pos = findAtt(globalAttsPos, att.getShortName());
-    else {
-      N3header.Vinfo vinfo = (N3header.Vinfo) v2.getSPobject();
-      pos = findAtt(vinfo.attsPos, att.getShortName());
-    }
-
-    raf.seek(pos);
-    int type = raf.readInt();
-    DataType have = getDataType(type);
-    DataType want = att.getDataType();
-    if (want == DataType.STRING)
-      want = DataType.CHAR;
-    if (want != have) {
-      throw new IllegalArgumentException(
-          "Update Attribute must have same type or original = " + have + " att = " + att);
-    }
-
-    if (type == 2) { // String
-      String s = att.getStringValue();
-      int org = raf.readInt();
-      int size = org + padding(org); // ok to use the padding
-      int max = Math.min(size, s.length()); // cant make any longer than size
-      if (max > org) { // adjust if its using the padding, but not if its shorter
-        raf.seek(pos + 4);
-        raf.writeInt(max);
-      }
-
-      byte[] b = new byte[size];
-      for (int i = 0; i < max; i++)
-        b[i] = (byte) s.charAt(i);
-      raf.write(b);
-
-    } else {
-      int nelems = raf.readInt();
-      int max = Math.min(nelems, att.getLength()); // cant make any longer
-      for (int j = 0; j < max; j++)
-        writeAttributeValue(att.getNumericValue(j));
-    }
-  }
-
-  private long findAtt(long start_pos, String want) throws IOException {
-    raf.seek(start_pos + 4);
-
-    int natts = raf.readInt();
-    for (int i = 0; i < natts; i++) {
-      String name = readString();
-      if (name.equals(want))
-        return raf.getFilePointer();
-
-      int type = raf.readInt();
-
-      if (type == 2) {
-        readString();
-      } else {
-        int nelems = raf.readInt();
-        DataType dtype = getDataType(type);
-        int[] shape = {nelems};
-        Array arr = Array.factory(dtype, shape);
-        IndexIterator ii = arr.getIndexIterator();
-        int nbytes = 0;
-        for (int j = 0; j < nelems; j++)
-          nbytes += readAttributeValue(dtype, ii);
-        skip(nbytes);
-      }
-    }
-
-    throw new IllegalArgumentException("no such attribute " + want);
-  }
-
-
 }
